[#28] Formats project ids in the template
Would be grateful on feedback on this to see what we think going
forward. Approach is to create an array from the path values of a
uri. Use a for loop and catch the last two items by knowing where
they will be.
Seems like there should be a better way.
This template is on a Licence view