[#16][#17][#18] Creates the HTML templates going forward
Much of this is just place holders.
This work 'breaks the back' of individual type views
and list views.

Further work could be done to template various sections using includes
but I'm unsure as to how we might pass variables to those includes
and slightly concerned that each view needs it's own minor customisations