Adds a new service -  map
The map on the projects list could be complemented or replaced by
a map at it's own page.
This has the advantage that page loads on the projects list should be
faster.
This change requires a docker rebuild to add the new service.
Tests for this page are minimal and juts test that the heading is there
I'm unsure how to better test for the map itself
I'm not sure how we would put this at e.g. project/map instead of at
/map