[#235] This might speed up country page load times
We have many queries that use optional clauses because we are not
sure about what is in the data. I think these generally slow things
down.
Because we know more about the data, we can remove some of these
because weknow for example that projects without a name do not need
to be shown.
The tweak to the sites query fixes a bug that inproves load time on
that query
Splitting the commodities and status queries out of the project query
I think improves things. Again with these queries we can now loose the
optional elements, and require to return just one value (Projects are
currently Oil and Gas OR Mining, and each only needs one status (even
if more are supplied they are duplicates)
I've split these out and then use a loop in the template to display
the correct ones against the correct project.
I think all of this is decreasing load times.
Ben, you might like to test and inspect what I have done to see if it
helps!