[#251] Removes more uneccessary html from the country page
There was a html comment that was being written out time and time again
And some more white space that could be trimmed.

Targetting the country page specifically here, so mexico may load a little
faster. In fact it probably does not load faster, but it halves the amount
of data sent to the browser, and I think that less data to the leaflet.js
map also improves things client side.