[#259] Applies suggested fix for payments table
I have tried this locally and I'm not sure it works as intended.

Putting the fix suggested in the issue into a branch so it's easier to
work with.

This could be a case where we write the test to demonstarte what we
should and should not see in the output.